--- src/core/commands.ts
@@ -1,8 +1,0 @@
-import pingCommand from "../modules/ping/commands/ping.command";
-
-/**
- * Commands array containing all commands to be registered.
- */
-const commands = [pingCommand];
-
-export default commands;
