+++ src/core/registry/slash-commands.ts
@@ -1,0 +1,8 @@
+import pingCommand from "../../modules/ping/commands/ping.command";
+
+/**
+ * Commands array containing all slash commands to be registered.
+ */
+const slashCommands = [pingCommand];
+
+export default slashCommands;
